update sidebar and editor tool items
