+++ src/containers/Editor/Tools.tsx
@@ -6,15 +6,18 @@ import {
 } from "react-icons/ai";
 import { FiDownload } from "react-icons/fi";
 import { HiOutlineSun, HiOutlineMoon } from "react-icons/hi";
+import { TiFlowMerge } from "react-icons/ti";
 import { MdCenterFocusWeak } from "react-icons/md";
+import { CgArrowsMergeAltH, CgArrowsShrinkH } from "react-icons/cg";
+import toast from "react-hot-toast";
+import { CanvasDirection } from "reaflow";
 import { SearchInput } from "src/components/SearchInput";
 import styled from "styled-components";
+import { getNextLayout } from "src/containers/Editor/LiveEditor/helpers";
 import useConfig from "src/hooks/store/useConfig";
 import shallow from "zustand/shallow";
 import { DownloadModal } from "../Modals/DownloadModal";
 import useStored from "src/hooks/store/useStored";
-import { TbSettings } from "react-icons/tb";
-import { Settings } from "./Settings";
 
 export const StyledTools = styled.div`
   display: flex;
@@ -46,14 +49,35 @@ const StyledToolElement = styled.button`
   }
 `;
 
+const Divider = styled.span`
+  height: 100%;
+  width: 1px;
+  background: ${({ theme }) => theme.BACKGROUND_MODIFIER_ACCENT};
+`;
+
+const StyledFlowIcon = styled(TiFlowMerge)<{ rotate: number }>`
+  transform: rotate(${({ rotate }) => `${rotate}deg`});
+`;
+
+function rotateLayout(layout: CanvasDirection) {
+  if (layout === "LEFT") return 90;
+  if (layout === "UP") return 180;
+  if (layout === "RIGHT") return 270;
+  return 360;
+}
+
 export const Tools: React.FC = () => {
-  const [settingsVisible, setSettingsVisible] = React.useState(false);
   const [isDownloadVisible, setDownloadVisible] = React.useState(false);
   const lightmode = useStored((state) => state.lightmode);
   const setLightTheme = useStored((state) => state.setLightTheme);
 
-  const [performanceMode, hideEditor] = useConfig(
-    (state) => [state.performanceMode, state.hideEditor],
+  const [performanceMode, layout, expand, hideEditor] = useConfig(
+    (state) => [
+      state.performanceMode,
+      state.layout,
+      state.expand,
+      state.hideEditor,
+    ],
     shallow
   );
 
@@ -64,17 +88,20 @@ export const Tools: React.FC = () => {
   const centerView = useConfig((state) => state.centerView);
   const toggleEditor = () => setConfig("hideEditor", !hideEditor);
   const toggleTheme = () => setLightTheme(!lightmode);
+  const toggleLayout = () => {
+    const nextLayout = getNextLayout(layout);
+    setConfig("layout", nextLayout);
+  };
+
+  const toggleExpandCollapse = () => {
+    setConfig("expand", !expand);
+    toast(`${expand ? "Collapsed" : "Expanded"} nodes.`);
+  };
 
   return (
     <StyledTools>
       <StyledToolElement aria-label="fullscreen" onClick={toggleEditor}>
         <AiOutlineFullscreen />
-      </StyledToolElement>
-      <StyledToolElement
-        aria-label="settings"
-        onClick={() => setSettingsVisible(true)}
-      >
-        <TbSettings />
       </StyledToolElement>
       <StyledToolElement aria-label="switch theme" onClick={toggleTheme}>
         {lightmode ? <HiOutlineMoon /> : <HiOutlineSun />}
@@ -89,6 +116,17 @@ export const Tools: React.FC = () => {
           <FiDownload />
         </StyledToolElement>
       )}
+      <Divider />
+      <StyledToolElement
+        aria-label={expand ? "shrink nodes" : "expand nodes"}
+        onClick={toggleExpandCollapse}
+      >
+        {expand ? <CgArrowsMergeAltH /> : <CgArrowsShrinkH />}
+      </StyledToolElement>
+      <StyledToolElement aria-label="toggle layout" onClick={toggleLayout}>
+        <StyledFlowIcon rotate={rotateLayout(layout)} />
+      </StyledToolElement>
+      <Divider />
       <StyledToolElement aria-label="center canvas" onClick={centerView}>
         <MdCenterFocusWeak />
       </StyledToolElement>
@@ -102,7 +140,6 @@ export const Tools: React.FC = () => {
         visible={isDownloadVisible}
         setVisible={setDownloadVisible}
       />
-      <Settings visible={settingsVisible} setVisible={setSettingsVisible} />
     </StyledTools>
   );
 };
